% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_func.R
\name{aggregate_func}
\alias{aggregate_func}
\title{Aggregation function for record-level data, prior to plot.}
\usage{
aggregate_func(mod_plot)
}
\arguments{
\item{mod_plot}{A data frame of values to be aggregated.  Expected columns, 'group', 'numerator' and 'denominator'}
}
\value{
Returns an aggregated data.frame of the same columns, but summed by 'group', with a relative risk 'rr' column added.
}
\description{
Internal function to aggregate record-level data for plotting as a funnel.
}
\keyword{internal}
