% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Funnel.R
\docType{package}
\name{CMFunnels}
\alias{CMFunnels}
\alias{CMFunnels-package}
\title{Funnel Plots of Incident reports}
\description{
This is a basic implementation of Speigelhalter's funnel plots for using wiht Chris Mainey's
PhD work, using aggregated incident reporting data from NRLS.  It expects a specific structure
of table and column names.
}
\author{
Chris Mainey
}
