% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_themes.R
\name{funnel_clean}
\alias{funnel_clean}
\title{A clean funnel plot theme}
\usage{
funnel_clean()
}
\value{
a list of ggplot theme items
}
\description{
A ggplot theme function for clean looking funnel plots.  Try funnel_grey if you like the old one.
}
\examples{
\dontrun{funnel_plot(theme=funnel_clean())}
}
\seealso{
funnel_grey
}
