% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_themes.R
\name{funnel_grey}
\alias{funnel_grey}
\title{A grey ggplot funnel theme}
\usage{
funnel_grey()
}
\value{
a list of ggplot theme items
}
\description{
A calssic ggplot theme function for funnel plots.  Try funnel_clean if you don't like the grey background.
}
\examples{
funnel_plot(theme=funnel_grey())
}
\seealso{
funnel_clean
}
