% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{limits.funnelplot}
\alias{limits.funnelplot}
\title{Extractor function for funnel plot limits}
\usage{
\method{limits}{funnelplot}(x)
}
\arguments{
\item{x}{object of class `funnelplot`}
}
\description{
Extractor function for funnel plot limits
}
