% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{new_funnel_plot}
\alias{new_funnel_plot}
\title{Constructor for new funnel plot object}
\usage{
new_funnel_plot(x = list())
}
\arguments{
\item{x}{List of objects to convert to class}
}
\description{
Constructor for new funnel plot object
}
