% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OD_adjust_func.R
\name{OD_adjust_func}
\alias{OD_adjust_func}
\title{Winsorisation and adjusted z-score function}
\usage{
OD_adjust_func(
  mod_plot_agg = mod_plot_agg,
  method = "SHMI",
  Winsorise_by = 0.1,
  multiplier = 1,
  bypass = FALSE
)
}
\arguments{
\item{mod_plot_agg}{Aggregated model input data}

\item{method}{Adjustment method, can take the value \"SHMI\" or \"CQC\". \"SHMI\" is default.}

\item{Winsorise_by}{The amount to winsorise\/truncate the distribution by, prior to transformation. 0.1 means 10\% (at each end).}

\item{multiplier}{The amount to multiply the standardised ratio by, default is 1.}

\item{bypass}{TRUE\/FALSE, whether to bypass adjustment (not yet in use in main function)}
}
\value{
A list with elements\: aggregated adjusted data fame, Phi (a numeric dispersion ratio), Tau2 (a numeric \"between\" standard error)'
}
\description{
Internal function to perform the winsorisation and adjustment of z-scores prior to plotting.
}
\keyword{internal}
