% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{outliers.funnelplot}
\alias{outliers.funnelplot}
\title{Extractor function for funnel plot outliers}
\usage{
\method{outliers}{funnelplot}(x)
}
\arguments{
\item{x}{object of class `funnelplot`}
}
\description{
Extractor function for funnel plot outliers
}
