% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Label outliers}
\usage{
outliers(mod_plot_agg, OD_adjust, Poisson_limits, limit)
}
\arguments{
\item{mod_plot_agg}{Aggregated data set for plotting}

\item{OD_adjust}{Logical for drawing OD limits, takes precedence over Poisson for outliers}

\item{Poisson_limits}{Logical for drawing Poisson limits}

\item{limit}{which limit to use.  Currently 95 or 99.}
}
\description{
Label outliers
}
