% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{phi.funnelplot}
\alias{phi.funnelplot}
\title{Extractor function for dispersion ratio, \eqn{\phi}, for funnel plots}
\usage{
\method{phi}{funnelplot}(x)
}
\arguments{
\item{x}{object of class `funnelplot`}
}
\description{
Extractor function for dispersion ratio, \eqn{\phi}, for funnel plots
}
