% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_adjustment.R
\name{phi_func}
\alias{phi_func}
\title{Calculate overdispersion ratio}
\usage{
phi_func(n, zscores)
}
\arguments{
\item{n}{Single numeric value for the count of the number of groups (and therefore z-scores)}

\item{zscores}{Vector of z-scores z-scores to be used.  Commonly, this would be 'winsorised' first to remove impact of extreme outliers.  SHMI truncates instead, but this simply reduced the n as well as the z-score.}
}
\value{
A numeric phi value
}
\description{
Internal function to perform the transformations for data types.
}
\keyword{internal}
