% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{plot.funnelplot}
\alias{plot.funnelplot}
\title{Extractor function for ggplot object of funnel plot}
\usage{
\method{plot}{funnelplot}(x, ...)
}
\arguments{
\item{x}{object of class `funnelplot`}

\item{...}{Other arguments}
}
\description{
Extractor function for ggplot object of funnel plot
}
