% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_limits.R
\name{poisson_limits}
\alias{poisson_limits}
\title{Poisson funnel limit calculation}
\usage{
poisson_limits(mod_plot_agg = mod_plot_agg, multiplier = 1, Target = Target)
}
\arguments{
\item{mod_plot_agg}{Aggregated model input data}

\item{multiplier}{Multiplier to adjust limits if reporting by a multiplier, e.g. per 1000.}

\item{Target}{supplied target value}
}
\value{
A data.frame of original, aggregated data plus transformed z-score (unadjusted for overdispersion)
}
\description{
Adds 95% and 99.8 % funnel limits from Poisson distribution, using exact Poisson limits.
}
