% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{print.funnelplot}
\alias{print.funnelplot}
\title{Print method for funnel plot}
\usage{
\method{print}{funnelplot}(x)
}
\arguments{
\item{x}{object of class `funnelplot`}
}
\description{
Print method for funnel plot
}
