% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{source_data.funnelplot}
\alias{source_data.funnelplot}
\title{Extractor function for funnel plot source data}
\usage{
\method{source_data}{funnelplot}(x)
}
\arguments{
\item{x}{object of class `funnelplot`}
}
\description{
Extractor function for funnel plot source data
}
