% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{source_data}
\alias{source_data}
\title{source data used to create Funnel plots}
\usage{
source_data(x)
}
\arguments{
\item{x}{object of class funnel plot}
}
\description{
Source data class for funnel plots
}
