% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{summary.funnelplot}
\alias{summary.funnelplot}
\title{Summary method for funnel plot}
\usage{
\method{summary}{funnelplot}(x)
}
\arguments{
\item{x}{object of class `funnelplot`}
}
\description{
Summary method for funnel plot
}
