% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{tau2.funnelplot}
\alias{tau2.funnelplot}
\title{Extractor function for between group variance, \eqn{\tau^2}, for funnel plots}
\usage{
\method{tau2}{funnelplot}(x)
}
\arguments{
\item{x}{object of class `funnelplot`}
}
\description{
Extractor function for between group variance, \eqn{\tau^2}, for funnel plots
}
