% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_limits.R
\name{transcribe_limits}
\alias{transcribe_limits}
\title{Transcribe names of control limits into test from values supplied to function}
\usage{
transcribe_limits(limits_z, Poisson_limits, OD_adjust)
}
\arguments{
\item{limits_z}{z-values for limits in question, e.g. 95\% are approx -1.96 & 1.96}

\item{Poisson_limits}{TRUE or FALSE are we drawing poisson control limits?}

\item{OD_adjust}{TRUE or FALSE are we drawing overdispersed poisson control limits?}
}
\value{
A vector of names for control limits
}
\description{
Transcribe names of control limits into test from values supplied to function
}
