% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OD_adjust_func.R
\name{transformed_zscore}
\alias{transformed_zscore}
\title{Transformation function for z-scoring}
\usage{
transformed_zscore(
  mod_plot_agg = mod_plot_agg,
  data_type = "SR",
  sr_method = "SHMI"
)
}
\arguments{
\item{mod_plot_agg}{Aggregated model input data}

\item{data_type}{Type of data for adjustment and plotting: Indirectly Standardised ratio (\"SR\"), proportion (\"PR\"), or ratio of counts (\"RC\").}

\item{sr_method}{Adjustment method, can take the value \"SHMI\" or \"CQC\". \"SHMI\" is default.}
}
\value{
A data.frame of original, aggreagated data plus transformed z-score (unadjusted for overdispersion)
}
\description{
Internal function to perform the transformations for data types.
}
\keyword{internal}
