% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_adjustment.R
\name{truncation}
\alias{truncation}
\title{Truncation function for NHSD method}
\usage{
truncation(mod_plot_agg = mod_plot_agg, trim_by = 0.1)
}
\arguments{
\item{mod_plot_agg}{Aggregated model input data}

\item{trim_by}{The amount to truncate the distribution by,
prior to transformation. 0.1 means 10\% (at each end).}
}
\value{
A data.frame with truncated z-scores added
}
\description{
Internal function to perform the truncation.
}
\keyword{internal}
