% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{tst_func}
\alias{tst_func}
\title{Test function}
\usage{
tst_func(.data, numerator, denominator, group)
}
\arguments{
\item{.data}{data frame}

\item{numerator}{tghe numearto}

\item{denominator}{denom}

\item{group}{the group}
}
\value{
a data frame stuck together
}
\description{
Test function
}
