% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{validate_funnel_plot}
\alias{validate_funnel_plot}
\title{Validator for new funnel plot object}
\usage{
validate_funnel_plot(funnelplot)
}
\arguments{
\item{funnelplot}{object of class funnelplot}
}
\description{
Validator for new funnel plot object
}
