% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OD_adjust_func.R
\name{winsorisation}
\alias{winsorisation}
\title{Winsorisation function}
\usage{
winsorisation(
  mod_plot_agg = mod_plot_agg,
  sr_method = "SHMI",
  Winsorise_by = 0.1
)
}
\arguments{
\item{mod_plot_agg}{Aggregated model input data}

\item{sr_method}{Adjustment method for standardised ratios, can take the value \"SHMI\" or \"CQC\". \"SHMI\" is default. Not relevant to PR or RC data types}

\item{Winsorise_by}{The amount to winsorise\/truncate the distribution by, prior to transformation. 0.1 means 10\% (at each end).}
}
\value{
A data.frame with winsorized z-scores returned added
}
\description{
Internal function to perform the winsorization or truncation.
}
\keyword{internal}
