% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnel_plot_function.R
\name{funnel_plot}
\alias{funnel_plot}
\title{Funnel Plots for Incident Reporting: A function for generating funnel plots or incident reporting data}
\usage{
funnel_plot(model, data, observed_name, OD_Tau2 = TRUE,
  Exact_limits = FALSE, label_outliers = TRUE, reportingyr, title,
  conditional = FALSE, ...)
}
\arguments{
\item{model}{The name of the model being plotted (lm, glm, lmer, glmer, mgcv::gam, glmmTMB). Alternatively, a vector of fitted values.}

\item{data}{dataframe of the source data for the plot.}

\item{observed_name}{name of the observed count column in dataframe.}

\item{OD_Tau2}{Draw overdispersed limit, using Speigelhalter's Tau2 (default=TRUE).}

\item{Exact_limits}{Draw exact limits based only on data points with no iterpolation. (default=FALSE).}

\item{label_outliers}{Add trust codes to outlier organisations on plot. (default=TRUE).}

\item{reportingyr}{Reporting year, if using more than one fiscal year, specify in format: '2015/16'.
Ignored if not specified.}

\item{title}{specify title for funnel plot, or default one is constructed from dataname.}

\item{conditional}{Predict lmer/glmer or GAMs with random effects as conditional? Default is FALSE, prediction is marginal
i.e. without random effect for organisation.}
}
\value{
a list of the funnel plot as GGPlot object, the limits table and the base table for plot.
}
\description{
This is a basic implementation of Speigelhalter's funnel plots for using wiht Chris Mainey's PhD work, using
aggregated incident reporting data from NRLS.  It expects a specific structure of table and column names.
}
\details{
There are a number of parameters for the input, with the assumption that

you will want smooth funnel limits plotted based on the DerSimmonian Laird Tau2 additive random
effects limits.

To use exact (with not limits outside of data range), set Exac_limits=TRUE. To use standard
2 & 3 sigma limits,

set OD_Tau2 to FALSE.  Outliers are marked based on Trust code.
}
\examples{
funnel_plot(Pois_count_1, incsub2a_st, 'incidents', '2015/16', 'Poisson model')
funnel_plot(Quasi_count_1, incsub2a_st, 'incidents', '2015/16', 'Quasi model')
funnel_plot(NB_count_1, incsub2a_st, 'incidents', OD_Tau2=FALSE, Exact_limits=FALSE, label_outliers=FALSE, '2015/16', 'NegBin model')
funnel_plot(Pois_count_1_rint, incsub2a_st, 'incidents', OD_Tau2=TRUE, Exact_limits=TRUE, label_outliers=TRUE,'2015/16', 'Poisson Rint model')
}
