% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_plot.R
\name{funnel_plot}
\alias{funnel_plot}
\title{Funnel Plots for Incident Reporting}
\usage{
funnel_plot(predictions, observed, group, title, label_outliers = TRUE,
  Poisson_limits = FALSE, OD_Tau2 = TRUE, method = "SHMI",
  Winsorize_by = 0.1, multiplier = 1, x_label = "Expected",
  y_label = "Standardised Ratio")
}
\arguments{
\item{predictions}{A vector of model predictions.  Used as denominator of the Y-axis and the scale of the x-axis}

\item{observed}{A vector of the observed value.  Used as numerator of the Y-axis}

\item{group}{A vector of group names or a factor.  Used to aggreagate and group points on plots}

\item{title}{Plot title}

\item{label_outliers}{Add group labels to outliers on plot. (default=TRUE)}

\item{Poisson_limits}{Draw exact limits based only on data points with no iterpolation. (default=FALSE)}

\item{OD_Tau2}{Draw overdispersed limits? Using Speigelhalter's (2012) Tau2 (default=TRUE)}

\item{method}{Either "CQC" or "SHMI" (default). There are a few methods for standardisation.  CQC/Spiegelhalter
uses a square root transformation and winsorizes by replaceing values, SHMI uses log transformation and winsorizes
by truncation. CQC method is default.}

\item{Winsorize_by}{Proportion of the distribution for winsorization. Default is 10 \% (0.1)}

\item{multiplier}{Scale relative risk and funnel by this factor. Default to 1, but 100 is used for HSMR}

\item{x_label}{Title for the funnel plot x-axis.  Usually expected deaths, readmissions, incidents etc.}

\item{y_label}{Title for the funnel plot y-axis.  Usually a standardised ratio.}
}
\value{
a list of the funnel plot as ggplot2 object, the limits table and the base table for plot.
}
\description{
This is an implementation of funnel plots described Spiegelhalter (2005).
There are several of parameters for the input, with the assumption that  you will want smooth,
 overdispersed, funnel limits plotted based on the DerSimmonian Laird \eqn{\tau^2} additive random
effects models.
}
\details{
Outliers are marked based on the grouping, controlled by `label_outliers` .
   Overdispersion can be factored in based on the methods in \href{https://rss.onlinelibrary.wiley.com/doi/full/10.1111/j.1467-985X.2011.01010.x}{Spiegelhalter et al (2012)}, set `OD_Tau2` to FALSE to suppress this.
   To use Poisson limits set `Poisson_limits=TRUE`. This uses 2 & 3 \eqn{\sigma} limits.
   It deliberatley avoids red-amber-green colouring, but you could extract this from the ggplot object and change manually if you like.
}
\examples{
\dontrun{
    a<-funnel_plot(my_preds, my_observed, "organisation", '2015/16', 'Poisson model')
    #Access the plot
    a[[3]]

    #Access the
    }


}
\seealso{
\href{https://rss.onlinelibrary.wiley.com/doi/full/10.1111/j.1467-985X.2011.01010.x}{Statistical methods for healthcare regulation: rating, screening and surveillance. Spiegelhalter et al (2012)}
   \href{https://onlinelibrary.wiley.com/doi/10.1002/sim.1970}{Funnel plots for comparing institutional performance. Spiegelhalter (2004)}
   \href{https://qualitysafety.bmj.com/content/14/5/347}{Handeling over-dispersion of performance indicators. Spiegelhalter (2005)}
}
