% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{summary.funnelplot}
\alias{summary.funnelplot}
\title{Summary method for funnel plot}
\usage{
\method{summary}{funnelplot}(object, ...)
}
\arguments{
\item{object}{object of class `funnelplot`}

\item{...}{Other arguments}
}
\description{
Summary method for funnel plot
}
