% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_adjustment.R
\name{winsorisation}
\alias{winsorisation}
\title{Winsorisation function}
\usage{
winsorisation(mod_plot_agg = mod_plot_agg, trim_by = 0.1)
}
\arguments{
\item{mod_plot_agg}{Aggregated model input data}

\item{trim_by}{The amount to Winsorise the distribution by, prior to transformation. 0.1 means 10\% (at each end).}
}
\value{
A data.frame with winsorised z-scores returned added
}
\description{
Internal function to perform the Winsorisation.
}
\keyword{internal}
